variables now use gtirb code style

+++ include/gtirb_pprinter/PrettyPrinter.hpp
@@ -475,7 +475,7 @@ protected:
 private:
   std::set<gtirb::Addr> functionEntry;
   std::set<gtirb::Addr> functionLastBlock;
-  std::map<const gtirb::Symbol*, uint16_t> AmbiguousSymbols;
+  std::map<const gtirb::Symbol*, size_t> AmbiguousSymbols;
   gtirb::Addr programCounter;
 
   std::optional<gtirb::Addr> CFIStartProc;
